requirements.txt

asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
defusedxml==0.7.1
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
idna==3.10
joblib==1.5.1
lxml==6.0.0
nltk==3.9.1
numpy==2.3.1
oauthlib==3.3.1
pycparser==2.22
PyJWT==2.9.0
PyPDF2==3.0.1
python-docx==1.2.0
python3-openid==3.2.0
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
scikit-learn==1.7.0
scipy==1.16.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
gunicorn
python-3.10.12
openai
whitenoise